Separate out the common functions in a separate module utilities.py. Also some general cleanup.
